Added sorting, did some fine tunings

+++ MainWindow.xaml.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.ObjectModel;
+using System.Data.Common;
 using System.Diagnostics;
 using System.Text;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Controls.Primitives;
 using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
@@ -33,9 +35,27 @@ namespace PersonDisplay
             _viewController.EditPanelEditable += PanelEditableEvent;
             _viewController.FillEditDataEvent += FillEditFieldsEvent;
             _viewController.ClearGrid += ClearGridEvent;
-            
+            _viewController.SetSortingDirectionEvent += SetSortingDirectionEvent;
             GridPanel.IsEnabled = false;
+            ClearInputs();
             WaitUntillReady();
+        }
+
+
+        private void SetSortingDirectionEvent(object sender, Tuple<string, bool> data)
+        {
+            DataGridColumn sortingColumn=null;
+            foreach (DataGridColumn column in PeopleGrid.Columns)
+            {
+                if(column.SortMemberPath==data.Item1)
+                    sortingColumn = column;
+            }
+            if (sortingColumn != null)
+            {
+                if(data.Item2) sortingColumn.SortDirection=System.ComponentModel.ListSortDirection.Ascending;
+                else sortingColumn.SortDirection = System.ComponentModel.ListSortDirection.Descending;
+            }
+            else throw new Exception("No such column!");
         }
 
         private async Task WaitUntillReady()
@@ -92,7 +112,9 @@ namespace PersonDisplay
         private void DeleteObjectEvent(object caller, RoutedEventArgs e)
         {
             Person person = (caller as Button).DataContext as Person;
-            _viewController.DeleteButtonClick(person);
+            MessageBoxResult reslut = MessageBox.Show("ARe you sure you want to delete " + person.Name + " " + person.Surname + "?", "Confirm deletion", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (reslut == MessageBoxResult.Yes)
+                _viewController.DeleteButtonClick(person);
         }
 
 
@@ -140,9 +162,10 @@ namespace PersonDisplay
             else GridPanel.Visibility = Visibility.Collapsed;
         }
 
-        private void SortTagEvent(object sender, DataGridSortingEventArgs e)
+        private void SortingEvent(object sender, DataGridSortingEventArgs e)
         {
-
+            e.Handled=true;
+            _viewController.UpdateSorter((e.Column).SortMemberPath);
         }
